nix/theme-manager/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "theme-manager"
version = "0.1.0"
description = "Daemon & CLI for managing themes"
readme = "README.md"
requires-python = ">=3.8"
dependencies = ["pyyaml", "pystray", "pillow"]

[project.scripts]
theme-manager = "theme_manager.daemon:main"
themectl = "themectl.main:main"


[tool.setuptools.package-dir]
"" = "src"